Addressed two TODOs.
